feat: create thread-safe cache structure
This is set up as a Rocket State which means it can be accessed from
everywhere in the api. It is thread-safe using Arc and Mutex types.

Right now this is simply a String to String store, but it could be
extended so that structured data could be kept in the store. The only
annoying thing would be that the regular 'flushing' of expired values
would need to handled separately for each Cache instance.

This is not a big problem but it deserves to be noted.