feat: handle toggle-connected event
Actually this is easy. The toggle-connected event can be emitted
from anywhere in DOM tree, with bubbles and composed set to true
it will always reach to the top of the tree. This means we just
have to register the handlers at the window level. The cool thing
is that multiple handlers can be registered for a same event, so
no conflict! It will just execute every handler registerd there.