feat: use cargo chef to optimize docker build
This is GREAT! cargo chef allows to separate dependency fetching
and building from binary building. This means that this step can
be staged by docker and does not have to be repeated each time a
source file is changed. The build time used to take around 5 min
and now it's more like 30s. This is a GREAT improvement.