fix: restart containers on failure
This is needed because depends_on does not wait for the setup to
be finished, it just waits for the container to be running. Then
the next one can fail if the first was not fast enough. The clean
way to do this would be to use the depends_on condition and set
it to 'service_healthy'. There would need to be a script checking
that the DB is indeed ready. This does not look too hard but the
'restart: on-failure' way is easier :D