feat: access Constants env variables in the front
I'm using cookies to do this. It's a bit heavy but it's worth it I
think. I have a single source of thruth now. And if the cookies are
flushed by the user there are some sensible defaults set in the
front. So all this changes nothing right now, but it will when we're
going use a different port for the API.

I could also have 'compiled' a config .json file to be used directly
in the front. Actually this might be better --' Hmm, I'll have to think
about this, see if it's doable. I think it is, I would just need to run
a script to create the file from environment in the Dockerfile (or maybe
there's a relevant compose field for this).