feat: forbid outside connections by default
In the compose file the service ports are bound to the HOST_IP_BIND
value if the variable exists. By default it is set to "127.0.0.1" so
that containers can only communicate via the loopback interface.
This effectively forbids network access, even local. This will be
useful in production so that the "true" service ports cannot be
accessed from the outside.